% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{data_deal}
\alias{data_deal}
\title{data_deal
data processing}
\usage{
data_deal(sc_rds, st_rds, plot_data)
}
\arguments{
\item{sc_rds}{Single-cell transcriptome data}

\item{st_rds}{Spatial transcriptome data}

\item{plot_data}{convolution data}
}
\description{
Preprocess the data to get the data we need later.
}
