% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{dataProcessing}
\alias{dataProcessing}
\title{dataProcessing
data processing}
\usage{
dataProcessing(
  sc_exp_data,
  st_exp_data,
  sc_celltype_data,
  loc_matrix,
  deconvolution_matrix
)
}
\arguments{
\item{sc_exp_data}{Single Cell Transcriptome Expression Matrix}

\item{st_exp_data}{Spatial transcriptome expression matrix}

\item{sc_celltype_data}{Cell type mete data}

\item{loc_matrix}{Spatial transcriptome coordinates information}

\item{deconvolution_matrix}{Spatial spot deconvolution results}
}
\description{
Preprocess the data to get the data we need later.
}
