% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{drim}
\alias{drim}
\title{planarian_main
main}
\usage{
drim(resolution = 4, thread = 7, colname)
}
\arguments{
\item{resolution}{The default program magnification is four times}

\item{thread}{The number of running threads, the default is 7}

\item{colname}{selected column name}

\item{conda_path}{Location of the conda}
}
\description{
If the environment is correct, you can run this program directly to get the result
}
