% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{env_python_set}
\alias{env_python_set}
\title{env_python_set}
\usage{
env_python_set(py_path)
}
\arguments{
\item{py_path}{path of the conda environment}
}
\description{
Set the path of the conda environment you use
Many python programs are called in our program,
so you need to set the path of the python interpreter,
and you can set it directly to the conda environment here.
}
