% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{env_test}
\alias{env_test}
\title{env_test
Detect environment dependencies of python}
\usage{
env_test()
}
\value{
bool TRUE or FALSE
}
\description{
You can use this function to detect if a package is missing from a dependent python environment.
}
