% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{planarian_main}
\alias{planarian_main}
\title{planarian_main
main}
\usage{
planarian_main(resolution = 4, colname)
}
\arguments{
\item{resolution}{The default program magnification is four times}

\item{colname}{selected column name}

\item{conda_path}{Location of the conda}
}
\description{
If the environment is correct, you can run this program directly to get the result
}
