% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{setParameters}
\alias{setParameters}
\title{setParameters
Hyperparameter setting}
\usage{
setParameters(resolution = 4, thread = 7, colname)
}
\arguments{
\item{resolution}{program magnification}

\item{colname}{selected column name}
}
\description{
Set Resolution and Cell Columns,resolution is the multiple of program amplification,
Cell Columns is the specified column name
}
