% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{testEnv}
\alias{testEnv}
\title{testEnv
Detect environment dependencies of python}
\usage{
testEnv()
}
\value{
bool TRUE or FALSE
}
\description{
You can use this function to detect if a package is missing from a dependent python environment.
}
