% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRIMC.R
\name{data_deal}
\alias{data_deal}
\title{data_deal
data processing}
\usage{
data_deal(sc_exp_data, st_exp_data, sc_celltype_data, loc_data, plot_data)
}
\arguments{
\item{sc_exp_data}{Single Cell Transcriptome Expression Matrix}

\item{st_exp_data}{Spatial transcriptome expression matrix}

\item{sc_celltype_data}{Cell type mete data}

\item{loc_data}{loc_data}

\item{plot_data}{convolution data}
}
\description{
Preprocess the data to get the data we need later.
}
